feat: UITableView에 대한 ProjectListViewModel 생성 #3
- 데이터 바인딩을 하는 Observable 타입 추가
- 배열을 안전하게 조회하기 위해 Collection extension 추가

+++ ProjectManager/ProjectManager/Sources/Domain/UseCases/MemoryUseCase.swift
@@ -15,13 +15,21 @@ extension MemoryUseCase: UseCase {
         }
     }
     
-    func update(_ project: Project, completion: ((Project?) -> Void)?) {
+    func update(_ project: Project?, completion: ((Project?) -> Void)?) {
+        guard let project = project else {
+            completion?(nil)
+            return
+        }
         repository.update(with: project) { project in
             completion?(project)
         }
     }
     
-    func delete(_ project: Project, completion: ((Project?) -> Void)?) {
+    func delete(_ project: Project?, completion: ((Project?) -> Void)?) {
+        guard let project = project else {
+            completion?(nil)
+            return
+        }
         repository.delete(project) { project in
             completion?(project)
         }
